+++ docs/getting-started/the-process.md
@@ -1,0 +1,1 @@
+# The process
