+++ docs/helpers/create-node.md
@@ -1,0 +1,1 @@
+# Create node
