+++ docs/helpers/functions.md
@@ -1,0 +1,1 @@
+# Functions
