+++ docs/helpers/general.md
@@ -1,0 +1,1 @@
+# General
