+++ docs/helpers/type-checking.md
@@ -1,0 +1,1 @@
+# Type checking
