+++ docs/index.md
@@ -1,0 +1,21 @@
+# JSCodeShaft 
+
+A collection of common helpers building on top of the excellent [jscodeshift](facebook/jscodeshift) and [recast](benjamn/recast). Its primary goal is to assist migrations and refactors or React codebases, with a focus on design systems.
+
+While it's used in production, it's a very slow WIP, so keep that in mind if you want to use it yourself.
+
+
+## Getting started
+1. [The process](getting-started/the-process.md)
+2. [Using this library](getting-started/using-this-library.md)
+3. [Adding your own helper](getting-started/adding-helpers.md)
+
+## Exposed helpers
+1. [Attributes](helpers/attributes.md)
+1. [CSS classnames](helpers/classnames.md)
+1. [Create node](helpers/create-node.md)
+1. [Functions](helpers/functions.md)
+1. [General](helpers/general.md)
+1. [Imports](helpers/imports.md)
+1. [JSX elements](helpers/jsx-elements.md)
+1. [Type Checking](helpers/type-checking.md)
