+++ lib/attributes/__tests__/getAttributes.js
@@ -168,7 +168,7 @@ describe('attributes::getAttributes()', () => {
     });
 
 
-    describe('emptyOnNoMatch=true', () => {
+    describe('allOnNoMatch=true', () => {
       it('attribute string: should return empty array on no match', () => {
         const absentProp = 'five';
 
@@ -218,7 +218,7 @@ describe('attributes::getAttributes()', () => {
     });
 
 
-    describe('emptyOnNoMatch=false', () => {
+    describe('allOnNoMatch', () => {
       it('attribute string: should return all attributes array on no match', () => {
         const absentProp = 'five';
 
@@ -232,7 +232,7 @@ describe('attributes::getAttributes()', () => {
 
           return getAttributes(noPropsElement, {
             wantedAttributes: absentProp,
-            emptyOnNoMatch: false,
+            allOnNoMatch: true,
           });
         };
 
@@ -263,7 +263,7 @@ describe('attributes::getAttributes()', () => {
 
           return getAttributes(noPropsElement, {
             wantedAttributes: absentProps,
-            emptyOnNoMatch: false,
+            allOnNoMatch: true,
           });
         };
