+++ lib/attributes/index.js
@@ -34,15 +34,15 @@ const toTruthyStringArray = (input) => {
  * Can be optionally customised via options - if filtered by name
  * using `wantedAttributes` parameter and no matches are found,
  * no attributes will be returned. To make it return all attributes
- * instead, override with `emptyOnNoMatch=false`.
+ * instead, override with `allOnNoMatch`.
  * @param   {Node}            node
  * @param   {string|string[]} [wantedAttributes]
- * @param   {boolean}         [emptyOnNoMatch=true]
+ * @param   {boolean}         [allOnNoMatch]
  * @returns {Node[]|undefined}
  */
 export const getAttributes = (node, {
   wantedAttributes,
-  emptyOnNoMatch = true,
+  allOnNoMatch,
 } = {}) => {
   let output;
 
@@ -61,7 +61,7 @@ export const getAttributes = (node, {
           }
         });
 
-        if (filteredAttrs.length > 0 || (filteredAttrs.length === 0 && emptyOnNoMatch)) {
+        if (filteredAttrs.length > 0 || (filteredAttrs.length === 0 && !allOnNoMatch)) {
           output = filteredAttrs;
         }
       }
