+++ lib/imports/index.js
@@ -61,7 +61,7 @@ export const getSpecifierNamesFromImport = (importNode, type) => {
  * This is useful if, for example, your codebase has multiple components
  * with the same name (internal or external) and you want to confirm you're
  * processing the right one.
- * @param   {object}    root  AST-ified file content
+ * @param   {Program} root  AST-ified file content
  * @returns {function(isImportedFromSourceArgs): boolean}
  */
 export const isImportedFromSource = root => (specifierName, importSource) => Boolean(
