+++ lib/index.js
@@ -4,6 +4,5 @@ export * from './attributes';
 export * from './jsxElements';
 export * from './imports';
 export * from './general';
-export * from './traversal';
 export * from './classNames';
 export * from './functions';
