+++ lib/jsxElements/index.js
@@ -1,3 +1,6 @@
+import {isText} from '../typeChecking';
+
+
 /**
  * @typedef {function}  elementArgs
  * @param   {string}    name        New element name
@@ -44,3 +47,21 @@ export const renameAllElements = (j, root) => (oldName, newName) => {
     })
   }
 };
+
+
+/**
+ * In some cases you need to get all children that are either elements,
+ * literals or expressions and ignore all text nodes that are just whitespace.
+ * @param   {Node[]} childrenNodes
+ * @returns {Node[]}
+ */
+export const getMeaningfulChildren = childrenNodes => childrenNodes?.filter(
+  childNode => {
+    const isTextNode = isText(childNode);
+
+    return !isTextNode || (
+      isTextNode && Boolean(childNode.value.trim().length)
+    )
+  }
+) || [];
+
