+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jscodeshaft",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "description": "Collection of more or less primitive helpers and abstractions for JSCodeShift, build for design system migrations and upgrades.",
   "repository": "https://github.com/criography/jscodeshaft",
   "main": "dist/index.js",
